WILDALERT/Wildanimal/requirements.txt
absl-py==2.0.0
altair==5.2.0
amqp==2.6.1
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
blinker==1.7.0
cachetools==5.3.2
catboost==1.2.2
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colormath==3.0.0
contourpy==1.2.0
cvzone==1.6.1
cycler==0.12.1
dataclasses==0.6
dj-database-url==2.1.0
Django==3.2
django-cleanup==8.0.0
django-crispy-forms==2.1
django-heroku==0.3.1
django-localflavor==4.0
django-timezone-field==4.2.3
djongo==1.2.38
dlib==19.24.2
dnspython==2.6.1
face-recognition==1.3.0
face_recognition_models==0.3.0
Faker==22.0.0
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.45.1
frozendict==2.4.0
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googlesearch-python==1.0.1
graphviz==0.20.1
grpcio==1.60.0
h5py==3.10.0
html5lib==1.1
ibm-db==3.2.1
idna==2.10
imageio==2.34.0
importlib-metadata==6.8.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
keras==2.15.0
kiwisolver==1.4.5
kombu==4.6.11
lazy_loader==0.3
libclang==16.0.6
lxml==5.1.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mediapipe==0.10.8
ml-dtypes==0.2.0
mpmath==1.3.0
multitasking==0.0.11
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
passlib==1.7.4
peewee==3.17.0
Pillow==10.1.0
playsound==1.2.2
plotly==5.18.0
protobuf==3.20.3
psutil==5.9.8
psycopg2==2.9.9
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
pymongo==4.6.2
pyparsing==3.1.1
pyserial==3.5
python-crontab==3.0.0
python-dateutil==2.8.2
python-stdnum==1.19
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
requests==2.25.1
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
soupsieve==2.5
sqlparse==0.2.4
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.2.12
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.1.20
urllib3==1.26.18
validators==0.22.0
vine==1.3.0
watchdog==3.0.0
webcolors==1.13
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
whois==0.9.13
wrapt==1.14.1
yfinance==0.2.35
zipp==3.17.0
